Cleaner file injector, added keyboard shortcuts
JS/CSS has been revamped. Instead of being included in the manifest, the background script first checks if the current tab is an instance page. If it is, it injects the JS/CSS files so that normal tabs are unaffected.

Keyboard shortcuts can now easily be bound to any button given the selector